# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Groove Atlas API — drum history and discovery
 * OpenAPI spec version: 0.2.0
 */
export interface HealthStatus {
  status: string;
}

export interface NotFound {
  message: string;
}

export interface Era {
  id: string;
  name: string;
  subtitle: string;
  years: string;
  description: string;
  color: string;
  characteristics: string[];
  keyDrummerIds: string[];
  iconicSongIds: string[];
}

export interface Drummer {
  id: string;
  name: string;
  born: number;
  died?: number | null;
  primaryEra: string;
  eras: string[];
  genres: string[];
  bands: string[];
  bio: string;
  signatureStyle: string;
  bpmMin: number;
  bpmMax: number;
  influence: string;
  iconicSongIds: string[];
}

export interface Song {
  id: string;
  title: string;
  artist: string;
  drummerId: string;
  year: number;
  eraId: string;
  genreIds: string[];
  tempo: number;
  feel: string;
  complexity: number;
  description: string;
  whyStudy: string;
  songsterrSlug?: string | null;
  lyricsSnippet?: string | null;
}

export type EraDetail = Era & {
  keyDrummers: Drummer[];
  iconicSongs: Song[];
};

export interface Genre {
  id: string;
  name: string;
  origin: string;
  lat: number;
  lng: number;
  era: string;
  description: string;
  color: string;
  characteristics: string[];
  keyDrummerIds: string[];
  iconicSongIds: string[];
}

export type GenreDetail = Genre & {
  keyDrummers: Drummer[];
  iconicSongs: Song[];
};

export type DrummerDetail = Drummer & {
  iconicSongs: Song[];
};

export type SongDetail = Song & ({
  drummer: Drummer | null;
});

export interface SearchResults {
  drummers: Drummer[];
  songs: Song[];
  genres: Genre[];
  eras: Era[];
}

export type ListDrummersParams = {
eraId?: string;
genreId?: string;
};

export type ListSongsParams = {
eraId?: string;
genreId?: string;
drummerId?: string;
};

export type SearchAllParams = {
q: string;
};