# lib/api-zod/src/generated/types/drummer.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Groove Atlas API — drum history and discovery
 * OpenAPI spec version: 0.2.0
 */

export interface Drummer {
  id: string;
  name: string;
  born: number;
  died?: number | null;
  primaryEra: string;
  eras: string[];
  genres: string[];
  bands: string[];
  bio: string;
  signatureStyle: string;
  bpmMin: number;
  bpmMax: number;
  influence: string;
  iconicSongIds: string[];
}