# lib/api-zod/src/generated/types/genre.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Groove Atlas API — drum history and discovery
 * OpenAPI spec version: 0.2.0
 */

export interface Genre {
  id: string;
  name: string;
  origin: string;
  lat: number;
  lng: number;
  era: string;
  description: string;
  color: string;
  characteristics: string[];
  keyDrummerIds: string[];
  iconicSongIds: string[];
}