# lib/api-zod/src/generated/types/genreDetail.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Groove Atlas API — drum history and discovery
 * OpenAPI spec version: 0.2.0
 */
import type { Drummer } from './drummer';
import type { Genre } from './genre';
import type { Song } from './song';

export type GenreDetail = Genre & {
  keyDrummers: Drummer[];
  iconicSongs: Song[];
};