# lib/api-zod/src/generated/types/notFound.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Groove Atlas API — drum history and discovery
 * OpenAPI spec version: 0.2.0
 */

export interface NotFound {
  message: string;
}