# lib/api-zod/src/generated/types/searchResults.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Groove Atlas API — drum history and discovery
 * OpenAPI spec version: 0.2.0
 */
import type { Drummer } from './drummer';
import type { Era } from './era';
import type { Genre } from './genre';
import type { Song } from './song';

export interface SearchResults {
  drummers: Drummer[];
  songs: Song[];
  genres: Genre[];
  eras: Era[];
}